The big component refactoring
