+++ components/builder/stack/builder/types.ts
@@ -1,0 +1,23 @@
+export const MODAL_TYPE_CERTIFICATE_TYPE = 'Certificate Type'
+export const MODAL_TYPE_IMMUNIZATION_STATUS = 'Immunization Status'
+export const MODAL_TYPE_VALUE = 'Value'
+export const MODAL_TYPE_VAR = 'Var'
+export const MODAL_TYPE_DATE = 'Date'
+export const MODAL_TYPE_COMPARE = 'Compare'
+export const MODAL_TYPE_COMPARE_DATE = 'Compare Date'
+export const MODAL_TYPE_COMPARE_IN = 'Compare In'
+export const MODAL_TYPE_IF = 'IF'
+export const MODAL_TYPE_AND = 'AND'
+
+export const MODAL_TYPE_CUSTOM_ALL = [MODAL_TYPE_CERTIFICATE_TYPE, MODAL_TYPE_IMMUNIZATION_STATUS]
+
+export const MODAL_TYPE_ALL = [
+  MODAL_TYPE_VALUE,
+  MODAL_TYPE_VAR,
+  MODAL_TYPE_DATE,
+  MODAL_TYPE_COMPARE,
+  MODAL_TYPE_COMPARE_DATE,
+  MODAL_TYPE_COMPARE_IN,
+  MODAL_TYPE_IF,
+  MODAL_TYPE_AND,
+]
