+++ translations/en/common.json
@@ -28,6 +28,7 @@
   "type": "Type",
   "dose": "Dose",
   "series": "Series",
+  "copy": "Copy",
   "export": "Export",
   "export.none": "Nothing to export",
 
@@ -146,6 +147,24 @@
   "builder.modal.vaccine": "Select Vaccine",
   "builder.modal.type": "Select Type",
   "builder.modal.export": "Export",
+  "builder.value": "Value",
+  "builder.var": "Var",
+  "builder.variable": "Variable",
+  "builder.date": "Date",
+  "builder.if": "IF",
+  "builder.then": "THEN",
+  "builder.else": "ELSE",
+  "builder.and": "AND",
+  "builder.in": "IN",
+  "builder.compare": "Compare",
+  "builder.comparedate": "Compare Date",
+  "builder.comparein": "Compare In",
+  "builder.additem": "Add Item",
+  "builder.certificatetype": "Certificate Type",
+  "builder.immunizationstatus": "Immunization Status",
+  "builder.select.component": "Select Component",
+  "builder.select.custom": "Custom",
+  "builder.select.default": "Default",
 
   "lang": {
     "en": "English",
